[fix] make max qty work with the website_decimal_qty module

+++ nfu_ecommerce/static/src/js/website_sale.js
@@ -15,8 +15,8 @@ odoo.define("nfu_ecommerce.website_max_qty", function (require) {
                 productIDs.push($(elem).find("span[data-product-id]").data("product-id"));
             });
             $input.data("update_change", true);
-            var max_qty = parseInt($input.closest("tr").find('input[name="max-qty"]').val(), 10);
-            var set_qty = parseInt($input.closest("tr").find('input[name!="max-qty"]').val(), 10);
+            var max_qty = parseFloat($input.closest("tr").find('input[name="max-qty"]').val(), 10);
+            var set_qty = parseFloat($input.closest("tr").find('input[name!="max-qty"]').val(), 10);
             if (max_qty < set_qty) {
                 max_qty = set_qty;
             }
@@ -30,7 +30,7 @@ odoo.define("nfu_ecommerce.website_max_qty", function (require) {
                 },
             }).then(function (data) {
                 $input.data("update_change", false);
-                var check_value = parseInt($input.val() || 0, 10);
+                var check_value = parseFloat($input.val() || 0, 10);
                 if (isNaN(check_value)) {
                     check_value = 1;
                 }
@@ -44,9 +44,9 @@ odoo.define("nfu_ecommerce.website_max_qty", function (require) {
                 }
                 $input.val(data.quantity);
                 $(".js_quantity[data-line-id=" + line_id + "]")
-                    .val(data.quantity)
-                    .text(data.quantity);
-
+                .val(data.quantity)
+                .text(data.quantity);
+                
                 wSaleUtils.updateCartNavBar(data);
                 wSaleUtils.showWarning(data.warning);
                 // Propagating the change to the express checkout forms
